[rename] 푸터 컴포넌트 파일명 변경 [isuue] #8
